# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Trading Bot Dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Signal {
  timestamp: string;
  symbol: string;
  ob_signal: string;
  ob_confidence: string;
  /** @nullable */
  bid_ask_ratio: number | null;
  /** @nullable */
  spread_pct?: number | null;
  /** @nullable */
  bid_wall_conc?: number | null;
  /** @nullable */
  ask_wall_conc?: number | null;
  /** @nullable */
  indicator_filter_passed?: string | null;
  /** @nullable */
  filter_reason?: string | null;
  /** @nullable */
  rsi?: number | null;
  /** @nullable */
  ema_fast?: number | null;
  /** @nullable */
  ema_slow?: number | null;
  /** @nullable */
  atr?: number | null;
  /** @nullable */
  volume_ratio?: number | null;
}

export interface Trade {
  timestamp: string;
  symbol: string;
  /** @nullable */
  signal?: string | null;
  /** @nullable */
  entry_price?: number | null;
  /** @nullable */
  stop_loss?: number | null;
  /** @nullable */
  take_profit?: number | null;
  /** @nullable */
  rr_ratio?: number | null;
  /** @nullable */
  quantity?: number | null;
  /** @nullable */
  notional?: number | null;
  /** @nullable */
  margin_used?: number | null;
  /** @nullable */
  margin_pct?: number | null;
  /** @nullable */
  total_cost?: number | null;
  /** @nullable */
  gross_profit?: number | null;
  /** @nullable */
  net_profit?: number | null;
  /** @nullable */
  dollar_risk_target?: number | null;
  /** @nullable */
  dollar_risk_actual?: number | null;
  /** @nullable */
  reduction_factor?: number | null;
  /** @nullable */
  adjustment_reason?: string | null;
  /** @nullable */
  sl_source?: string | null;
  /** @nullable */
  sl_atr_dist?: number | null;
  /** @nullable */
  sl_ob_dist?: number | null;
  /** @nullable */
  sl_distance?: number | null;
  /** @nullable */
  atr?: number | null;
  /** @nullable */
  ob_confidence?: string | null;
  /** @nullable */
  ob_imbalance?: string | null;
  /** @nullable */
  allowed: string | null;
  /** @nullable */
  rejection_reason?: string | null;
}

export interface Position {
  timestamp: string;
  symbol: string;
  signal: string;
  /** @nullable */
  entry_price: number | null;
  /** @nullable */
  close_price: number | null;
  /** @nullable */
  stop_loss_at_close?: number | null;
  /** @nullable */
  take_profit?: number | null;
  /** @nullable */
  quantity?: number | null;
  /** @nullable */
  gross_pnl?: number | null;
  /** @nullable */
  net_pnl?: number | null;
  /** @nullable */
  total_cost?: number | null;
  /** @nullable */
  trailing_active?: string | null;
  /** @nullable */
  exit_reason?: string | null;
  /** @nullable */
  exit_type?: string | null;
}

export interface BotSummary {
  total_signals: number;
  long_signals?: number;
  short_signals?: number;
  neutral_signals?: number;
  total_trades_evaluated: number;
  trades_approved?: number;
  trades_rejected?: number;
  win_count?: number;
  loss_count?: number;
  win_rate: number;
  total_gross_pnl?: number;
  total_net_pnl: number;
  avg_win?: number;
  avg_loss?: number;
  avg_rr?: number;
  max_drawdown?: number;
  exit_tp_count?: number;
  exit_sl_count?: number;
  exit_ts_count?: number;
  open_positions_count?: number;
  paper_balance: number;
  current_balance?: number;
}

export interface SymbolStat {
  symbol: string;
  trades: number;
  win_count: number;
  loss_count: number;
  win_rate: number;
  total_net_pnl: number;
  avg_net_pnl?: number;
}

export interface SizingRecord {
  timestamp: string;
  symbol: string;
  /** @nullable */
  signal: string | null;
  /** @nullable */
  dollar_risk_target?: number | null;
  /** @nullable */
  dollar_risk_actual?: number | null;
  /** @nullable */
  reduction_factor?: number | null;
  /** @nullable */
  margin_used?: number | null;
  /** @nullable */
  margin_pct?: number | null;
  /** @nullable */
  adjustment_reason?: string | null;
  /** @nullable */
  notional?: number | null;
  /** @nullable */
  net_profit?: number | null;
}

export interface SummaryBlock {
  block: number;
  start_cycle: number;
  end_cycle: number;
  timestamp: string;
  trades_opened: number;
  trades_closed: number;
  win_count?: number;
  loss_count?: number;
  win_rate?: number;
  expectancy?: number;
  profit_factor?: number;
  max_drawdown?: number;
  total_net_pnl?: number;
  mark_to_market_closes?: number;
}

export interface Alert {
  type: string;
  severity: string;
  blockIds: number[];
  metric: string;
  values: number[];
  triggered: boolean;
  /**
     * @minimum 0
     * @maximum 1
     */
  confidence: number;
}

export interface AlertsResponse {
  timestamp: number;
  alerts: Alert[];
}

export interface PnlPoint {
  timestamp: string;
  symbol: string;
  net_pnl: number;
  cumulative_net_pnl: number;
  /** @nullable */
  exit_type?: string | null;
}

export type GetSignalsParams = {
limit?: number;
};

export type GetTradesParams = {
limit?: number;
};

export type GetSizingParams = {
limit?: number;
};