# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Trading Bot Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bot/signals:
    get:
      operationId: getSignals
      tags: [bot]
      summary: Recent OB scan signals
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 100 }
      responses:
        "200":
          description: List of recent signal scans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Signal"
  /bot/trades:
    get:
      operationId: getTrades
      tags: [bot]
      summary: Recent trade decisions (approved + rejected)
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 100 }
      responses:
        "200":
          description: List of trade decisions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trade"
...
  /bot/sizing:
    get:
      operationId: getSizing
      tags: [bot]
      summary: Adaptive sizing history for approved trades
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 100 }
      responses:
        "200":
          description: List of sizing records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SizingRecord"
...
components:
  schemas:
...
    Signal:
      type: object
      required: [timestamp, symbol, ob_signal, ob_confidence, bid_ask_ratio]
      properties:
        timestamp:               { type: string }
        symbol:                  { type: string }
        ob_signal:               { type: string }
        ob_confidence:           { type: string }
        bid_ask_ratio:           { type: ["number", "null"] }
        spread_pct:              { type: ["number", "null"] }
        bid_wall_conc:           { type: ["number", "null"] }
        ask_wall_conc:           { type: ["number", "null"] }
        indicator_filter_passed: { type: ["string", "null"] }
        filter_reason:           { type: ["string", "null"] }
        rsi:                     { type: ["number", "null"] }
        ema_fast:                { type: ["number", "null"] }
        ema_slow:                { type: ["number", "null"] }
        atr:                     { type: ["number", "null"] }
        volume_ratio:            { type: ["number", "null"] }
    Trade:
      type: object
      required: [timestamp, symbol, allowed]
      properties:
        timestamp:            { type: string }
        symbol:               { type: string }
        signal:               { type: ["string", "null"] }
        entry_price:          { type: ["number", "null"] }
        stop_loss:            { type: ["number", "null"] }
        take_profit:          { type: ["number", "null"] }
        rr_ratio:             { type: ["number", "null"] }
        quantity:             { type: ["number", "null"] }
        notional:             { type: ["number", "null"] }
        margin_used:          { type: ["number", "null"] }
        margin_pct:           { type: ["number", "null"] }
        total_cost:           { type: ["number", "null"] }
        gross_profit:         { type: ["number", "null"] }
        net_profit:           { type: ["number", "null"] }
        dollar_risk_target:   { type: ["number", "null"] }
        dollar_risk_actual:   { type: ["number", "null"] }
        reduction_factor:     { type: ["number", "null"] }
        adjustment_reason:    { type: ["string", "null"] }
        sl_source:            { type: ["string", "null"] }
        sl_atr_dist:          { type: ["number", "null"] }
        sl_ob_dist:           { type: ["number", "null"] }
        sl_distance:          { type: ["number", "null"] }
        atr:                  { type: ["number", "null"] }
        ob_confidence:        { type: ["string", "null"] }
        ob_imbalance:         { type: ["string", "null"] }
        allowed:              { type: ["string", "null"] }
        rejection_reason:     { type: ["string", "null"] }
...
    SizingRecord:
      type: object
      required: [timestamp, symbol, signal]
      properties:
        timestamp:          { type: string }
        symbol:             { type: string }
        signal:             { type: ["string", "null"] }
        dollar_risk_target: { type: ["number", "null"] }
        dollar_risk_actual: { type: ["number", "null"] }
        reduction_factor:   { type: ["number", "null"] }
        margin_used:        { type: ["number", "null"] }
        margin_pct:         { type: ["number", "null"] }
        adjustment_reason:  { type: ["string", "null"] }
        notional:           { type: ["number", "null"] }
        net_profit:         { type: ["number", "null"] }
...